backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Yansu — Inspeção Óptica Inteligente (Backend API)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.48",
    "asyncpg>=0.30.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.20",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "anomalib>=1.1.0",
    "opencv-python-headless>=4.8.0",
]


